+++ public/webpack.dev.js
@@ -2,7 +2,8 @@ var merge = require('webpack-merge')
   , common = require('./webpack.common.js');
 
 module.exports = merge(common, {
-  devtool: 'inline-source-map',
+  mode: 'development',
+  devtool: 'inline-module-source-map',
   devServer: {
     contentBase: './dist',
     inline: true,