+++ public/webpack.prod.js
@@ -1,9 +1,6 @@
 var merge = require('webpack-merge')
-  , common = require('./webpack.common.js')
-  , UglifyJSPlugin = require('uglifyjs-webpack-plugin');
+  , common = require('./webpack.common.js');
 
 module.exports = merge(common, {
-  plugins: [
-    new UglifyJSPlugin()
-  ]
+  mode: 'production'
 });